+++ allennlp/data/__init__.py
@@ -1,3 +1,4 @@
+from allennlp.data.dataloader import DataLoader, allennlp_collate
 from allennlp.data.dataset_readers.dataset_reader import DatasetReader
 from allennlp.data.fields.field import DataArray, Field
 from allennlp.data.fields.text_field import TextFieldTensors
